Add files via upload
Reference:
Hamid Mohammadi, Vahid Morovati, Abd-Elrahman Korayem, Emad Poshtan, Roozbeh Dargazany
Constitutive modeling of elastomers during photo-and thermo-oxidative aging
Polymer Degradation and Stability

Summary:
n this study, we proposed a micro-mechanical model to describe the effects of temperature and UV radiation on the constitutive behavior of cross-linked polymeric systems such as elastomers. Specifically, the focus of this model is to predict the effects of photo-oxidative aging on the mechanical response of cross-linked systems and their inelastic responses such as permanent set and Mullins effect. With regard to experimental studies on the nature of chemical reactions during aging, chain scission and bond creation should be considered as two competing phenomena that happens simultaneously, though at different rates. Consequently, cross-link density across the matrix might increase or decrease in the course of photo-oxidative aging, depending on the rate of reactions. The rates are functions of temperature, UV radiation intensity and time. In view of the competition between these two mechanisms, the elastomer can be damaged differently during aging which may lead to a harder or softer matrix, or even in more complicated cases, change between hardening and softening in due time. Therefore, strain energy of the polymer matrix should be calculated with regard to time, temperature and UV radiation intensity. Using this experimental analysis, we proposed a new micro-mechanical model. Since separating the effects of thermo-oxidation and photo-oxidation seemed impossible, the proposed model is a generalized micro-mechanical model for both thermo-oxidative and photo-oxidative aging. Therefore, the model has two sets of parameters. One that its components only are a function of temperature and another that are a function of both temperature and UV radiation intensity.